Fix signedness warning in glib/gspawn-win32.c
glib/gspawn-win32.c: In function 'read_helper_report':
glib/gspawn-win32.c:329:16: warning: comparison of integer expressions of different signedness: 'gint' {aka 'int'} and 'long long unsigned int'
   while (bytes < sizeof(gintptr)*2)
                ^
glib/gspawn-win32.c:366:13: warning: comparison of integer expressions of different signedness: 'gint' {aka 'int'} and 'long long unsigned int'
   if (bytes < sizeof(gintptr)*2)
             ^

+++ glib/gspawn-win32.c
@@ -324,7 +324,7 @@ read_helper_report (int      fd,
 		    gintptr  report[2],
 		    GError **error)
 {
-  gint bytes = 0;
+  gsize bytes = 0;
   
   while (bytes < sizeof(gintptr)*2)
     {
